CloudFunctions/enriched_table/requirements.txt
functions-framework==3.5.0
requests==2.31.0
google-cloud==0.34.0
google-cloud-storage==2.13.0
google-cloud-bigquery==3.13.0